Refactor BetterString class into separate module
The class BetterString has been relocated from __init__.py and moved to
a newly created separate module BetterString.py. This reorganization
improves modularity and makes the code easier to maintain.

Additionally, a new optional stacking string has been added to allow you
 to sequence several nodes in a row and concatenate their results!

Neat!
