+++ __init__.py
@@ -1,29 +1,13 @@
-class BetterString:
+from .BetterString import *
 
-    def __init__(self):
-        pass
-
-    @classmethod
-    def INPUT_TYPES(s):
-        return {
-            "required": {
-                "string": ("STRING", {"multiline": True}),
-            },
-        }
-
-    FUNCTION = "event"
-    RETURN_TYPES = ("STRING",)
-
-    CATEGORY = "Better Things 💡"
-
-    def event(self, string):
-        return (string, )
-
+# from .BetterPath import *
 
 NODE_CLASS_MAPPINGS = {
-    "BetterString": BetterString
+    "BetterString": BetterString,
+    # "BetterPath": BetterPath,
 }
 
 NODE_DISPLAY_NAME_MAPPINGS = {
-    "BetterString": "Better String"
+    "BetterString": "Better Multiline String",
+    # "BetterPath": "Better Path Selector",
 }
